Refactor linear attention backend with KernelDispatcher abstraction
Split GDN and KDA backends from the monolithic hybrid_linear_attn_backend.py
into separate files under layers/attention/linear/, adding a KernelDispatcher
abstraction layer so new kernel implementations can be added without modifying
backend logic.

- Add LinearAttnKernelBackend enum (auto/triton/cutedsl) and --linear-attn-kernel-backend server arg
- Add LinearAttnKernelCore ABC with decode/extend/target_verify methods
- Add concrete kernel wrappers: TritonGDNKernel, CuteDSLGDNKernel, TritonKDAKernel
- Add GDNKernelDispatcher and KDAKernelDispatcher for per-mode kernel resolution
- Move GDNAttnBackend to linear/gdn_backend.py, refactored to use dispatcher
- Move KimiLinearAttnBackend to linear/kda_backend.py as KDAAttnBackend (alias kept)
- Backward-compatible lazy re-exports via __getattr__ in hybrid_linear_attn_backend.py
