+++ python/sglang/srt/layers/attention/linear/kda_kernel_dispatcher.py
@@ -1,0 +1,84 @@
+import torch
+
+from sglang.srt.layers.attention.linear.kernels.triton_kda import TritonKDAKernel
+from sglang.srt.layers.attention.linear.utils import LinearAttnKernelBackend
+from sglang.srt.utils.common import rank0_log
+
+
+class KDAKernelDispatcher:
+    """Dispatches KDA kernel calls to the appropriate backend per mode.
+
+    Currently only Triton is supported for KDA. The dispatcher is structured
+    to allow future kernel additions without modifying the backend.
+    """
+
+    def __init__(self, backend: LinearAttnKernelBackend):
+        triton_kernel = TritonKDAKernel()
+
+        if backend.is_auto() or backend.is_triton():
+            self._decode_kernel = triton_kernel
+            self._extend_kernel = triton_kernel
+        else:
+            raise ValueError(
+                f"Unsupported KDA kernel backend: {backend}. "
+                "KDA currently only supports 'auto' and 'triton'."
+            )
+
+        rank0_log(
+            f"KDA kernel dispatcher: decode={self._decode_kernel.__class__.__name__}, "
+            f"extend={self._extend_kernel.__class__.__name__}"
+        )
+
+    def decode(
+        self,
+        q: torch.Tensor,
+        k: torch.Tensor,
+        v: torch.Tensor,
+        a: torch.Tensor,
+        b: torch.Tensor,
+        *,
+        A_log: torch.Tensor,
+        dt_bias: torch.Tensor,
+        ssm_states: torch.Tensor,
+        cache_indices: torch.Tensor,
+        query_start_loc: torch.Tensor,
+        **kwargs,
+    ) -> torch.Tensor:
+        return self._decode_kernel.decode(
+            q,
+            k,
+            v,
+            a,
+            b,
+            A_log=A_log,
+            dt_bias=dt_bias,
+            ssm_states=ssm_states,
+            cache_indices=cache_indices,
+            query_start_loc=query_start_loc,
+            **kwargs,
+        )
+
+    def extend(
+        self,
+        q: torch.Tensor,
+        k: torch.Tensor,
+        v: torch.Tensor,
+        g: torch.Tensor,
+        beta: torch.Tensor,
+        *,
+        ssm_states: torch.Tensor,
+        cache_indices: torch.Tensor,
+        query_start_loc: torch.Tensor,
+        **kwargs,
+    ) -> torch.Tensor:
+        return self._extend_kernel.extend(
+            q,
+            k,
+            v,
+            g,
+            beta,
+            ssm_states=ssm_states,
+            cache_indices=cache_indices,
+            query_start_loc=query_start_loc,
+            **kwargs,
+        )
