+++ python/sglang/srt/layers/attention/linear/kernels/__init__.py
@@ -1,0 +1,7 @@
+from sglang.srt.layers.attention.linear.kernels.triton_gdn import TritonGDNKernel
+from sglang.srt.layers.attention.linear.kernels.triton_kda import TritonKDAKernel
+
+__all__ = [
+    "TritonGDNKernel",
+    "TritonKDAKernel",
+]
